backend/requirements.txt
# Python dependencies for backend

fastapi
uvicorn
ultralytics
opencv-python
pillow
numpy
